+++ R/tabularaster-package.r
@@ -7,9 +7,20 @@
 #'  \code{\link{as_tibble}}
 #'  \code{\link{as_tibble}} \tab convert raster data to data frame form, with control over output and form of dimension/coordinate columns \cr
 #'  \code{\link{cellnumbers}} \tab extract a data frame of query identifiers and cell,pixel index numbers \cr
-#'  \code{\link{extentFromCells}} \tab build an Extent object from cell numbers \cr
+#'  \code{\link{extentFromCells}} \tab (now incorporated into raster itself) \cr
 #'  \code{\link{index_extent}} \tab build an extent in row column form, as opposed to coordinate value form \cr
 #'  }
+NULL
+
+#' Re-exports from raster
+#' 
+#' These functions are imported and then re-exported.
+#' 
+#' See \code{raster::\link[raster]{extentFromCells}} for details.
+#' @name raster-exports 
+#' @aliases extentFromCells 
+#' @importFrom raster extentFromCells
+#' @export extentFromCells
 NULL
 
 #' Sea surface temperature data.